requirements.txt
youtube-transcript-api
llama-index-llms-gemini
streamlit
#json
#llama-index google-generativeai
#regex